moved test files to relevant directories
Removed XLayout.java which didn't seem to serve any purpose
