+++ flying-saucer-examples/src/main/java/DelegatingReplacedElementFactory.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  * }}}
  */
-package com.github.danfickle.flyingsaucer.swing.test;
+
 
 import org.jsoup.nodes.Element;
 import org.xhtmlrenderer.extend.ReplacedElement;
