+++ flying-saucer-examples/src/main/java/SwingImageReplacer.java
@@ -1,4 +1,4 @@
-package com.github.danfickle.flyingsaucer.swing.test;
+
 
 import org.jsoup.nodes.Element;
 import org.xhtmlrenderer.extend.ReplacedElement;
