+++ flying-saucer-swing/src/test/java/com/github/danfickle/flyingsaucer/swing/test/HamletSpeedTest.java
